Updated workflow

+++ Sources/Vulkan/VulkanRender.cpp
@@ -186,7 +186,7 @@ void VulkanRender::RecreateSwapchain()
 void VulkanRender::UpdateUniformBuffers()
 {
     vk::Extent2D extent = mSwapchain->GetExtent();
-    float aspectRatio = static_cast<float>(extent.width) / extent.height;
+    float aspectRatio = static_cast<float>(extent.width) / static_cast<float>(extent.height);
 
     Core::UniformBufferObject ubo;
     ubo.view = mScene.GetCamera()->Transform();
